Move pressure-monitor test to its own non-parallel collection
The Theory in SKObjectTest asserts on process-wide GC.CollectionCount
deltas. With xUnit's default per-class parallelization, allocations
from other test classes running concurrently can pollute the counts
and flake the monitorEnabled=false case (observed locally: spurious
Gen0 collections during the measurement window).

Move the Theory to SKNativeMemoryPressureMonitorTest and put it in a
[CollectionDefinition(DisableParallelization = true)] collection so it
runs without any concurrent test class. SKObjectTest's other tests
keep their default parallel behavior.
